Fix: be more lenient when registering duplicate functions
