+++ examples/sushi/macros/macros.py
@@ -1,14 +1,6 @@
+from macros.utils import between  # type: ignore
+
 from sqlmesh import macro
-
-
-@macro()
-def add_one(evaluator, column):
-    return f"{column} + 1"
-
-
-@macro()
-def between(evaluator, value, start, end):
-    return value.between(start, end)
 
 
 @macro()
