+++ examples/sushi/macros/utils.py
@@ -1,0 +1,11 @@
+from sqlmesh import macro
+
+
+@macro()
+def add_one(evaluator, column):
+    return f"{column} + 1"
+
+
+@macro()
+def between(evaluator, value, start, end):
+    return value.between(start, end)
